pyproject.toml
[project]
name = "google-python-api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-api-python-client==2.187.0",
    "python-dateutil>=2.9.0.post0",
    "python-dotenv==1.2.1",
    "rich>=13.9.4",
]
